cleaned up util package a bit
