+++ src/main/java/me/borawski/hcf/listener/ChatListener.java
@@ -1,15 +1,24 @@
 package me.borawski.hcf.listener;
 
+import java.util.function.Consumer;
+
+import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.AsyncPlayerChatEvent;
 
+import com.massivecraft.factions.entity.Faction;
+import com.massivecraft.factions.entity.MPlayer;
+
 import me.borawski.hcf.Core;
+import me.borawski.hcf.session.FactionSession;
+import me.borawski.hcf.session.FactionSessionHandler;
 import me.borawski.hcf.session.Session;
 import me.borawski.hcf.session.SessionHandler;
-import me.borawski.hcf.util.ChatProcessor;
 import me.borawski.hcf.util.ChatUtils;
+import mkremins.fanciful.FancyMessage;
 
 public class ChatListener implements Listener {
 
@@ -28,7 +37,33 @@ public class ChatListener implements Listener {
             s.sendMessage(ChatColor.DARK_GRAY + "-----------------------------------------------------");
             return;
         }
-        ChatProcessor.handleChat(event.getMessage(), event.getPlayer());
+
+        String msg = event.getMessage();
+        Player player = event.getPlayer();
+
+        Bukkit.getOnlinePlayers().stream().forEach(new Consumer<Player>() {
+            @Override
+            public void accept(Player players) {
+                Session s = SessionHandler.getSession(player);
+                Faction f = MPlayer.get(player).getFaction();
+
+                if (f == null) {
+                    new FancyMessage(s.getRank().getPrefix() + " " + player.getName() + ": " + s.getRank().getColor() + msg).tooltip(new String[] {
+                            ChatColor.RED + "User is not in a faction"
+                    }).send(players);
+                    return;
+                }
+
+                FactionSession fSession = FactionSessionHandler.getFactionSession(f.getName());
+
+                new FancyMessage(s.getRank().getPrefix() + " " + player.getName() + ": " + s.getRank().getColor() + msg).tooltip(new String[] {
+                        ChatColor.DARK_RED + "" + ChatColor.BOLD + "FACTION INFO",
+                        ChatColor.GRAY + "Name: " + ChatColor.YELLOW + "" + fSession.getName(),
+                        ChatColor.GRAY + "Members: " + ChatColor.YELLOW + "" + f.getMPlayers().size(),
+                        ChatColor.GRAY + "Trophy Points: " + ChatColor.YELLOW + "" + fSession.getTrophies()
+                }).send(players);
+            }
+        });
     }
 
 }
