--- src/main/java/me/borawski/hcf/util/BarUtil.java
@@ -1,74 +1,0 @@
-package me.borawski.hcf.util;
-
-import net.minecraft.server.v1_11_R1.*;
-import org.bukkit.ChatColor;
-import org.bukkit.craftbukkit.v1_11_R1.entity.CraftPlayer;
-import org.bukkit.entity.Player;
-
-import java.lang.reflect.Field;
-
-/**
- * Created by Ethan on 3/12/2017.
- */
-public class BarUtil {
-
-    public static void sendActionBar(Player p, String msg) {
-        IChatBaseComponent cbc = IChatBaseComponent.ChatSerializer.a("{\"text\": \"" + msg + "\"}");
-        PacketPlayOutChat ppoc = new PacketPlayOutChat(cbc, (byte) 2);
-        ((CraftPlayer) p).getHandle().playerConnection.sendPacket(ppoc);
-    }
-
-    public static void sendTitle(Player player, Integer fadeIn, Integer stay, Integer fadeOut, String title, String subtitle) {
-        PlayerConnection connection = ((CraftPlayer) player).getHandle().playerConnection;
-        PacketPlayOutTitle packetPlayOutTimes = new PacketPlayOutTitle(PacketPlayOutTitle.EnumTitleAction.TIMES, null, fadeIn.intValue(), stay.intValue(), fadeOut.intValue());
-        connection.sendPacket(packetPlayOutTimes);
-        IChatBaseComponent titleMain;
-        PacketPlayOutTitle packetPlayOutTitle;
-        if (subtitle != null) {
-            subtitle = subtitle.replaceAll("%player%", player.getDisplayName());
-            subtitle = ChatColor.translateAlternateColorCodes('&', subtitle);
-            titleMain = IChatBaseComponent.ChatSerializer.a("{\"text\": \"" + subtitle + "\"}");
-            packetPlayOutTitle = new PacketPlayOutTitle(PacketPlayOutTitle.EnumTitleAction.SUBTITLE, titleMain);
-            connection.sendPacket(packetPlayOutTitle);
-        }
-
-        if (title != null) {
-            title = title.replaceAll("%player%", player.getDisplayName());
-            title = ChatColor.translateAlternateColorCodes('&', title);
-            titleMain = IChatBaseComponent.ChatSerializer.a("{\"text\": \"" + title + "\"}");
-            packetPlayOutTitle = new PacketPlayOutTitle(PacketPlayOutTitle.EnumTitleAction.TITLE, titleMain);
-            connection.sendPacket(packetPlayOutTitle);
-        }
-    }
-
-    public static void sendTabTitle(Player player, String header, String footer) {
-        if (header == null) {
-            header = "";
-        }
-
-        header = ChatColor.translateAlternateColorCodes('&', header);
-        if (footer == null) {
-            footer = "";
-        }
-
-        footer = ChatColor.translateAlternateColorCodes('&', footer);
-        header = header.replaceAll("%player%", player.getDisplayName());
-        footer = footer.replaceAll("%player%", player.getDisplayName());
-        PlayerConnection connection = ((CraftPlayer) player).getHandle().playerConnection;
-        IChatBaseComponent tabTitle = IChatBaseComponent.ChatSerializer.a("{\"text\": \"" + header + "\"}");
-        IChatBaseComponent tabFoot = IChatBaseComponent.ChatSerializer.a("{\"text\": \"" + footer + "\"}");
-        PacketPlayOutPlayerListHeaderFooter headerPacket = new PacketPlayOutPlayerListHeaderFooter(tabTitle);
-
-        try {
-            Field e = headerPacket.getClass().getDeclaredField("b");
-            e.setAccessible(true);
-            e.set(headerPacket, tabFoot);
-        } catch (Exception var11) {
-            var11.printStackTrace();
-        } finally {
-            connection.sendPacket(headerPacket);
-        }
-
-    }
-
-}
