--- src/main/java/me/borawski/hcf/util/KickUtil.java
@@ -1,23 +1,0 @@
-package me.borawski.hcf.util;
-
-import org.bukkit.ChatColor;
-
-/**
- * Created by Ethan on 3/12/2017.
- */
-public class KickUtil {
-
-    public static String getKick(Reason reason) {
-        String primary = (ChatColor.RED + "" + ChatColor.BOLD);
-        return primary + "Your account has been kicked!\n\n" +
-                "Reason: " + ChatColor.WHITE + reason.name() + primary + "\n";
-    }
-
-    public enum Reason {
-        DATA_CHANGE,
-        MANUAL,
-        ERROR,
-        WARNING,
-    }
-
-}
