--- src/main/java/me/borawski/hcf/util/PunishmentUtil.java
@@ -1,7 +1,0 @@
-package me.borawski.hcf.util;
-
-/**
- * Created by Ethan on 3/12/2017.
- */
-public class PunishmentUtil {
-}
